[check] break up __init__.py
