+++ python_modules/dagster/dagster/_check/builder.py
@@ -1,0 +1,290 @@
+import collections.abc
+import sys
+from contextlib import contextmanager
+from contextvars import ContextVar
+from typing import (
+    Any,
+    Callable,
+    ForwardRef,
+    Mapping,
+    NamedTuple,
+    Optional,
+    Tuple,
+    Type,
+    Union,
+    get_args,
+    get_origin,
+)
+
+from typing_extensions import Annotated
+
+from .functions import CheckError, TypeOrTupleOfTypes, failed
+
+try:
+    # this type only exists in python 3.10+
+    from types import UnionType  # type: ignore
+except ImportError:
+    UnionType = Union
+
+NoneType = type(None)
+
+_contextual_ns: ContextVar[Mapping[str, Type]] = ContextVar("_contextual_ns", default={})
+
+
+class EvalContext(NamedTuple):
+    """Utility class for managing references to global and local namespaces.
+
+    These namespaces are passed to ForwardRef._evaluate to resolve the actual
+    type from a string.
+    """
+
+    global_ns: dict
+    local_ns: dict
+
+    @staticmethod
+    def capture_from_frame(
+        depth: int,
+        *,
+        add_to_local_ns: Optional[Mapping[str, Any]] = None,
+    ) -> "EvalContext":
+        """Capture the global and local namespaces via the stack frame.
+
+        Args:
+            depth: which stack frame to reference, with depth 0 being the callsite.
+            add_to_local_ns: A mapping of additional values to update the local namespace with.
+
+        """
+        ctx_frame = sys._getframe(depth + 1)  # noqa # surprisingly not costly
+
+        # copy to not mess up frame data
+        global_ns = ctx_frame.f_globals.copy()
+        local_ns = ctx_frame.f_locals.copy()
+
+        if add_to_local_ns:
+            local_ns.update(add_to_local_ns)
+
+        return EvalContext(
+            global_ns=global_ns,
+            local_ns=local_ns,
+        )
+
+    @staticmethod
+    @contextmanager
+    def contextual_namespace(ns: Mapping[str, Type]):
+        token = _contextual_ns.set(ns)
+        try:
+            yield
+        finally:
+            _contextual_ns.reset(token)
+
+    def update_from_frame(self, depth: int):
+        # Update the global and local namespaces with symbols from the target frame
+        ctx_frame = sys._getframe(depth + 1)  # noqa # surprisingly not costly
+        self.global_ns.update(ctx_frame.f_globals)
+        self.local_ns.update(ctx_frame.f_locals)
+
+    def get_merged_ns(self):
+        return {
+            **_contextual_ns.get(),
+            **self.global_ns,
+            **self.local_ns,
+        }
+
+    def eval_forward_ref(self, ref: ForwardRef) -> Optional[Type]:
+        try:
+            if sys.version_info <= (3, 9):
+                return ref._evaluate(self.get_merged_ns(), {})  # noqa
+            else:
+                return ref._evaluate(self.get_merged_ns(), {}, frozenset())  # noqa
+        except NameError as e:
+            raise CheckError(
+                f"Unable to resolve {ref}, could not map string name to actual type."
+            ) from e
+
+    def compile_fn(self, body: str, fn_name: str) -> Callable:
+        local_ns = {}
+        exec(
+            body,
+            self.get_merged_ns(),
+            local_ns,
+        )
+        return local_ns[fn_name]
+
+
+def _coerce_type(
+    ttype: Optional[TypeOrTupleOfTypes],
+    eval_ctx: Optional[EvalContext],
+) -> Optional[TypeOrTupleOfTypes]:
+    # coerce input type in to the type we want to pass to the check call
+
+    # Any type translates to passing None for the of_type argument
+    if ttype is Any:
+        return None
+
+    # assume naked strings should be ForwardRefs
+    if isinstance(ttype, str):
+        if eval_ctx is None:
+            failed(
+                f"Can not generate check calls from string {ttype} (assumed ForwardRef) without EvalContext"
+            )
+        return eval_ctx.eval_forward_ref(ForwardRef(ttype))
+    if isinstance(ttype, ForwardRef):
+        if eval_ctx is None:
+            failed(f"Can not evaluate ForwardRef {ttype} without passing in EvalContext")
+        return eval_ctx.eval_forward_ref(ttype)
+
+    # Unions should become a tuple of types to pass to the of_type argument
+    # ultimately used as second arg in isinstance(target, tuple_of_types)
+    if get_origin(ttype) in (UnionType, Union):
+        union_types = get_args(ttype)
+        coerced_types = []
+        for t in union_types:
+            # coerce all the inner types
+            coerced = _coerce_type(t, eval_ctx)
+            if coerced is None or isinstance(coerced, tuple):
+                failed(f"Unable to coerce Union member {t} in {ttype}")
+            coerced_types.append(coerced)
+
+        return tuple(t for t in coerced_types)
+
+    return ttype
+
+
+def _container_pair_args(
+    args: Tuple[Type, ...], eval_ctx
+) -> Tuple[Optional[TypeOrTupleOfTypes], Optional[TypeOrTupleOfTypes]]:
+    # process tuple of types as if its two arguments to a container type
+
+    if len(args) == 2:
+        return _coerce_type(args[0], eval_ctx), _coerce_type(args[1], eval_ctx)
+
+    return None, None
+
+
+def _container_single_arg(
+    args: Tuple[Type, ...], eval_ctx: Optional[EvalContext]
+) -> Optional[TypeOrTupleOfTypes]:
+    # process tuple of types as if its the single argument to a container type
+
+    if len(args) == 1:
+        return _coerce_type(args[0], eval_ctx)
+
+    return None
+
+
+def _name(target: Optional[TypeOrTupleOfTypes]) -> str:
+    # turn a type or tuple of types in to its string representation for printing
+
+    if target is None:
+        return "None"
+
+    if isinstance(target, tuple):
+        return f"({', '.join(tup_type.__name__ if tup_type is not NoneType else 'check.NoneType' for tup_type in target)})"
+
+    return target.__name__
+
+
+def build_check_call_str(
+    ttype: Type,
+    name: str,
+    eval_ctx: Optional[EvalContext],
+) -> str:
+    # assumes this module is in global/local scope as check
+    origin = get_origin(ttype)
+    args = get_args(ttype)
+
+    # scalars
+    if origin is None:
+        if ttype is str:
+            return f'{name} if isinstance({name}, str) else check.str_param({name}, "{name}")'
+        elif ttype is float:
+            return f'{name} if isinstance({name}, float) else check.float_param({name}, "{name}")'
+        elif ttype is int:
+            return f'{name} if isinstance({name}, int) else check.int_param({name}, "{name}")'
+        elif ttype is bool:
+            return f'{name} if isinstance({name}, bool) else check.bool_param({name}, "{name}")'
+        elif ttype is Any:
+            return name  # no-op
+
+        # fallback to inst
+        inst_type = _coerce_type(ttype, eval_ctx)
+        if inst_type:
+            it = _name(inst_type)
+            return (
+                f'{name} if isinstance({name}, {it}) else check.inst_param({name}, "{name}", {it})'
+            )
+        else:
+            return name  # no-op
+    else:
+        # 3.9+: origin is Annotated, 3.8: origin == args[0]
+        if (origin is Annotated and args) or (len(args) == 1 and args[0] == origin):
+            return build_check_call_str(args[0], f"{name}", eval_ctx)
+
+        pair_left, pair_right = _container_pair_args(args, eval_ctx)
+        single = _container_single_arg(args, eval_ctx)
+
+        # containers
+        if origin is list:
+            return f'check.list_param({name}, "{name}", {_name(single)})'
+        elif origin is dict:
+            return f'check.dict_param({name}, "{name}", {_name(pair_left)}, {_name(pair_right)})'
+        elif origin is set:
+            return f'check.set_param({name}, "{name}", {_name(single)})'
+        elif origin is collections.abc.Sequence:
+            return f'check.sequence_param({name}, "{name}", {_name(single)})'
+        elif origin is collections.abc.Iterable:
+            return f'check.iterable_param({name}, "{name}", {_name(single)})'
+        elif origin is collections.abc.Mapping:
+            return f'check.mapping_param({name}, "{name}", {_name(pair_left)}, {_name(pair_right)})'
+        elif origin is collections.abc.Set:
+            return f'check.set_param({name}, "{name}", {_name(single)})'
+        elif origin in (UnionType, Union):
+            # optional
+            if pair_right is type(None):
+                inner_origin = get_origin(pair_left)
+                # optional scalar
+                if inner_origin is None:
+                    if pair_left is str:
+                        return f'{name} if {name} is None or isinstance({name}, str) else check.opt_str_param({name}, "{name}")'
+                    elif pair_left is float:
+                        return f'{name} if {name} is None or isinstance({name}, float) else check.opt_float_param({name}, "{name}")'
+                    elif pair_left is int:
+                        return f'{name} if {name} is None or isinstance({name}, int) else check.opt_int_param({name}, "{name}")'
+                    elif pair_left is bool:
+                        return f'{name} if {name} is None or isinstance({name}, bool) else check.opt_bool_param({name}, "{name}")'
+
+                    # fallback to opt_inst
+                    inst_type = _coerce_type(pair_left, eval_ctx)
+                    it = _name(inst_type)
+                    if inst_type:
+                        return f'{name} if {name} is None or isinstance({name}, {it}) else check.opt_inst_param({name}, "{name}", {it})'
+                    else:
+                        return name  # no-op
+
+                # optional container
+                else:
+                    inner_args = get_args(pair_left)
+                    inner_pair_left, inner_pair_right = _container_pair_args(inner_args, eval_ctx)
+                    inner_single = _container_single_arg(inner_args, eval_ctx)
+                    if inner_origin is list:
+                        return f'check.opt_nullable_list_param({name}, "{name}", {_name(inner_single)})'
+                    elif inner_origin is dict:
+                        return f'check.opt_nullable_dict_param({name}, "{name}", {_name(inner_pair_left)}, {_name(inner_pair_right)})'
+                    elif inner_origin is set:
+                        return (
+                            f'check.opt_nullable_set_param({name}, "{name}", {_name(inner_single)})'
+                        )
+                    elif inner_origin is collections.abc.Sequence:
+                        return f'check.opt_nullable_sequence_param({name}, "{name}", {_name(inner_single)})'
+                    elif inner_origin is collections.abc.Iterable:
+                        return f'check.opt_nullable_iterable_param({name}, "{name}", {_name(inner_single)})'
+                    elif inner_origin is collections.abc.Mapping:
+                        return f'check.opt_nullable_mapping_param({name}, "{name}", {_name(inner_pair_left)}, {_name(inner_pair_right)})'
+            # union
+            else:
+                tuple_types = _coerce_type(ttype, eval_ctx)
+                if tuple_types is not None:
+                    tt_name = _name(tuple_types)
+                    return f'{name} if isinstance({name}, {tt_name}) else check.inst_param({name}, "{name}", {tt_name})'
+
+        failed(f"Unhandled {ttype}")
